Fix some more uninentional implicit integer conversions

+++ toxcore/Messenger.c
@@ -337,7 +337,7 @@ int32_t m_create_group_connection(Messenger *m, GC_Chat *chat)
     }
 
     int onion_friend_number = friend_conn_get_onion_friendnum(connection);
-    Onion_Friend *onion_friend = onion_get_friend(m->onion_c, onion_friend_number);
+    Onion_Friend *onion_friend = onion_get_friend(m->onion_c, (uint16_t)onion_friend_number);
 
     onion_friend_set_gc_public_key(onion_friend, get_chat_id(chat->chat_public_key));
     onion_friend_set_gc_data(onion_friend, nullptr, -1);
@@ -1062,7 +1062,7 @@ int send_conference_invite_packet(const Messenger *m, int32_t friendnumber, cons
  *  return 0 on success
  *  return -1 on failure
  */
-int send_group_invite_packet(const Messenger *m, uint32_t friendnumber, const uint8_t *data, size_t length)
+int send_group_invite_packet(const Messenger *m, uint32_t friendnumber, const uint8_t *data, uint16_t length)
 {
     if (write_cryptpacket_id(m, friendnumber, PACKET_ID_INVITE_GROUPCHAT, data, length, 0)) {
         return 0;
